implement, basic test, move dtype normalization stuff to new module
